Remove tests in CI
Our automated tests check for broken links. Unfortunately, we're seeing more and more errors failing unrelated pull requests. Often, this is not due to a page no longer existing, but a site now blocking non-individual IPs.

This PR removes the CI check, but leaves it in place for locals (where we can confirm things work from an unblocked IP).